feat(rhino): added additional view props to converter
