+++ Objects/Converters/ConverterRhinoGh/ConverterRhinoGh.Utils.cs
@@ -1,17 +1,68 @@
-﻿using Objects;
+﻿using Grasshopper.Kernel.Types;
+using Objects.Geometry;
+using Objects.Primitive;
+using Objects.Other;
 using Rhino;
+using Rhino.Geometry;
+using Rhino.DocObjects;
 using Speckle.Core.Kits;
 using Speckle.Core.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Objects.Converter.RhinoGh
 {
   public partial class ConverterRhinoGh
   {
+    private RenderMaterial GetMaterial(RhinoObject o)
+    {
+      var material = o.GetMaterial(true);
+      var renderMaterial = new RenderMaterial();
+
+      // If it's a default material use the display color.
+      if (!material.HasId)
+      {
+        renderMaterial.diffuse = o.Attributes.DrawColor(Doc).ToArgb();
+        return renderMaterial;
+      }
+
+      // Otherwise, extract what properties we can. 
+      renderMaterial.name = material.Name;
+      renderMaterial.diffuse = material.DiffuseColor.ToArgb();
+      renderMaterial.emissive = material.EmissionColor.ToArgb();
+      renderMaterial.opacity = 1 - material.Transparency;
+      renderMaterial.metalness = material.Reflectivity;
+
+      if (material.Name.ToLower().Contains("glass") && renderMaterial.opacity == 0)
+      {
+        renderMaterial.opacity = 0.3;
+      }
+
+      return renderMaterial;
+    }
+
+    private string GetSchema(RhinoObject obj, out string[] args)
+    {
+      args = null;
+
+      // user string has format "DirectShape{[family], [type]}" if it is a directshape conversion
+      // otherwise, it is just the schema type name
+      string schema = obj.Attributes.GetUserString(SpeckleSchemaKey);
+
+      if (schema == null)
+        return null;
+
+      string[] parsedSchema = schema.Split(new char[] { '(', ')' }, StringSplitOptions.RemoveEmptyEntries);
+      if (parsedSchema.Length > 2) // there is incorrect formatting in the schema string!
+        return null;
+      if (parsedSchema.Length == 2)
+        args = parsedSchema[1].Split(new char[] { ',' }, StringSplitOptions.RemoveEmptyEntries).Select(o => o.Trim()).ToArray();
+      return parsedSchema[0].Trim();
+    }
+
+    #region Units
+
     /// <summary>
     /// Computes the Speckle Units of the current Document. The Rhino document is passed as a reference, so it will always be up to date.
     /// </summary>    
@@ -27,7 +78,6 @@ namespace Objects.Converter.RhinoGh
       var f = Units.GetConversionFactor(units, ModelUnits);
       return value * f;
     }
-
 
     private string UnitToSpeckle(UnitSystem us)
     {
@@ -90,11 +140,7 @@ namespace Objects.Converter.RhinoGh
         default:
           throw new System.Exception("The current Unit System is unsupported.");
       }
-
-
     }
-
-
-
+    #endregion
   }
 }
