+++ Objects/Converters/ConverterRhinoGh/ConverterRhinoGh.cs
@@ -1,9 +1,9 @@
 ﻿using Grasshopper.Kernel.Types;
 using Objects.Geometry;
 using Objects.Primitive;
+using Objects.Other;
 using Rhino;
 using Rhino.Geometry;
-using Rhino.Display;
 using Rhino.DocObjects;
 using Speckle.Core.Kits;
 using Speckle.Core.Models;
@@ -30,7 +30,6 @@ using RH = Rhino.Geometry;
 
 using Surface = Objects.Geometry.Surface;
 using Vector = Objects.Geometry.Vector;
-using Objects.Other;
 
 namespace Objects.Converter.RhinoGh
 {
@@ -468,53 +467,6 @@ namespace Objects.Converter.RhinoGh
         default:
           return false;
       }
-    }
-
-    private RenderMaterial GetMaterial(RhinoObject o)
-    {
-      var material = o.GetMaterial(true);
-      var renderMaterial = new RenderMaterial();
-
-      // If it's a default material use the display color.
-      if (!material.HasId)
-      {
-        renderMaterial.diffuse = o.Attributes.DrawColor(Doc).ToArgb();
-        return renderMaterial;
-      }
-
-      // Otherwise, extract what properties we can. 
-      renderMaterial.name = material.Name;
-      renderMaterial.diffuse = material.DiffuseColor.ToArgb();
-      renderMaterial.emissive = material.EmissionColor.ToArgb();
-
-      renderMaterial.opacity = 1 - material.Transparency;
-      renderMaterial.metalness = material.Reflectivity;
-
-      if (material.Name.ToLower().Contains("glass") && renderMaterial.opacity == 0)
-      {
-        renderMaterial.opacity = 0.3;
-      }
-
-      return renderMaterial;
-    }
-
-    private string GetSchema(RhinoObject obj, out string[] args)
-    {
-      args = null;
-
-      // user string has format "DirectShape{[family], [type]}" if it is a directshape conversion
-      // otherwise, it is just the schema type name
-      string schema = obj.Attributes.GetUserString(SpeckleSchemaKey);
-
-      if (schema == null)
-        return null;
-
-      string[] parsedSchema = schema.Split(new char[] { '(', ')' }, StringSplitOptions.RemoveEmptyEntries);
-      if (parsedSchema.Length > 2) // there is incorrect formatting in the schema string!
-        return null;
-      if (parsedSchema.Length == 2)
-        args = parsedSchema[1].Split(new char[] { ',' }, StringSplitOptions.RemoveEmptyEntries).Select(o => o.Trim()).ToArray();
-      return parsedSchema[0].Trim();
     }
   }
 }
