+++ Objects/Objects/BuiltElements/View.cs
@@ -19,7 +19,6 @@ namespace Objects.BuiltElements
     public Point origin { get; set; }
     public Vector upDirection { get; set; }
     public Vector forwardDirection { get; set; }
-    //public double zoom { get; set; }
     public bool isOrthogonal { get; set; } = false;
 
     public View3D() { }
